Parametrize custom config page
